new file:   compile.sh 	modified:   examples/csections_db/CH4.py 	modified:   examples/csections_db/He.py 	modified:   examples/postprocess_two_stream.py 	modified:   examples/presets/two_stream_grids.py 	modified:   examples/run_two_stream.py 	modified:   examples/util/frames.py 	modified:   src/_default/run_poisson_eq.cxx 	modified:   src/_ltplib/_ltplib.cxx 	modified:   src/_ltplib/csection_set_cfg.cxx 	modified:   src/_ltplib/csection_set_utils.cxx 	modified:   src/_ltplib/def_pstore.cxx 	modified:   src/_ltplib/def_pstore.hxx 	modified:   src/_ltplib/io_strings.hxx 	modified:   src/api_backend.hxx

+++ src/_ltplib/def_pstore.hxx
@@ -15,7 +15,7 @@ py::array_t<f32, py::array::c_style> \
 parts_output_t;
 
 typedef \
-std::function<size_t(u8, const parts_input_t&)> \
+std::function<u32(u8, const parts_input_t&)> \
 inject_fn_t;
 
 typedef \
